migrate  to Filament 4 default class structure
