+++ src/Filament/Admin/Resources/Permissions/Pages/ManagePermissions.php
