--- src/Filament/Admin/Resources/ProgramResource.php
@@ -1,110 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources;
-
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Infolists\Components\TextEntry;
-use Filament\Infolists\Infolist;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Support\Str;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\ProgramResource\Pages;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\ProgramResource\RelationManagers;
-
-class ProgramResource extends Resource
-{
-    protected static ?string $navigationIcon = 'heroicon-o-building-library';
-
-    public static function getModel(): string
-    {
-        return config('filament-team-management.models.program');
-    }
-
-    public static function shouldRegisterNavigation(): bool
-    {
-        return config('filament-team-management.use_programs');
-    }
-
-    public static function getNavigationGroup(): string
-    {
-        if (config('filament-team-management.use_programs')) {
-            return 'Programs, ' . Str::ucfirst(Str::plural(config('filament-team-management.names.team'))) . ' and Users';
-        } else {
-            return Str::ucfirst(Str::plural(config('filament-team-management.names.team'))) . ' and Users';
-        }
-    }
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\Section::make('Program Details')
-                    ->schema([
-                        Forms\Components\TextInput::make('name')
-                            ->required()
-                            ->maxLength(255),
-                        Forms\Components\Textarea::make('description'),
-                        Forms\Components\Textarea::make('note'),
-                    ]),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('name')
-                    ->sortable()
-                    ->searchable(),
-                Tables\Columns\TextColumn::make('description')
-                    ->sortable()
-                    ->searchable(),
-                Tables\Columns\TextColumn::make('note')
-                    ->sortable()
-                    ->searchable(),
-                Tables\Columns\TextColumn::make('teams_count')
-                    ->label('# ' . Str::ucfirst(Str::plural(config('filament-team-management.names.team'))))
-                    ->counts('teams')
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('users_count')
-                    ->label('# Users')
-                    ->counts('users')
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('invites_count')
-                    ->label('# Invites')
-                    ->counts('invites')
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('created_at')
-                    ->sortable(),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            RelationManagers\TeamsRelationManager::class,
-            RelationManagers\UsersRelationManager::class,
-            RelationManagers\InvitesRelationManager::class,
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ListPrograms::route('/'),
-            'create' => Pages\CreateProgram::route('/create'),
-            'edit' => Pages\EditProgram::route('/{record}/edit'),
-            'view' => Pages\ViewProgram::route('/{record}'),
-        ];
-    }
-
-    public static function infolist(Infolist $infolist): Infolist
-    {
-        return $infolist
-            ->schema([
-                TextEntry::make('description')->hiddenLabel(),
-            ]);
-    }
-}
