--- src/Filament/Admin/Resources/ProgramResource/Pages/CreateProgram.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\ProgramResource\Pages;
-
-use Filament\Resources\Pages\CreateRecord;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\ProgramResource;
-
-class CreateProgram extends CreateRecord
-{
-    protected static string $resource = ProgramResource::class;
-}
