--- src/Filament/Admin/Resources/ProgramResource/Pages/InteractsWithRecordOfType.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\ProgramResource\Pages;
-
-trait InteractsWithRecordOfType {}
