+++ src/Filament/Admin/Resources/Programs/Schemas/ProgramForm.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\Programs\Schemas;
+
+use Filament\Forms\Components\DateTimePicker;
+use Filament\Forms\Components\TextInput;
+use Filament\Schemas\Schema;
+
+class ProgramForm
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->components([
+                TextInput::make('name')
+                    ->required(),
+                TextInput::make('email')
+                    ->label('Email address')
+                    ->email()
+                    ->required(),
+                DateTimePicker::make('email_verified_at'),
+                TextInput::make('password')
+                    ->password()
+                    ->required(),
+            ]);
+    }
+}
