--- src/Filament/Admin/Resources/TeamResource.php
@@ -1,107 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources;
-
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Infolists\Components\TextEntry;
-use Filament\Infolists\Infolist;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Support\Str;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource\Pages;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource\RelationManagers\InvitesRelationManager;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource\RelationManagers\UsersRelationManager;
-
-// filament-odk-link package related code are commented as some applications may not require ODK functionalities.
-// Please uncomment those code if filament-odk-link package is required and added to main repo.
-
-class TeamResource extends Resource
-{
-    protected static ?string $navigationIcon = 'heroicon-o-building-office-2';
-
-    public static function getModel(): string
-    {
-        return config('filament-team-management.models.team');
-    }
-
-    public static function getNavigationGroup(): string
-    {
-        if (config('filament-team-management.use_programs')) {
-            return 'Programs, ' . Str::ucfirst(Str::plural(config('filament-team-management.names.team'))) . ' and Users';
-        } else {
-            return Str::ucfirst(Str::plural(config('filament-team-management.names.team'))) . ' and Users';
-        }
-    }
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\Section::make(Str::ucfirst(config('filament-team-management.names.team')) . ' Details')
-                    ->schema([
-                        Forms\Components\TextInput::make('name')
-                            ->required()
-                            ->maxLength(255),
-                        Forms\Components\Textarea::make('description'),
-                    ]),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('name')
-                    ->searchable()
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('programs.name')
-                    ->searchable()
-                    ->badge()
-                    ->color('success')
-                    ->visible(config('filament-team-management.use_programs')),
-                Tables\Columns\TextColumn::make('users_count')
-                    ->label('# Users')
-                    ->counts('users')
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('invites_count')
-                    ->label('# Invites')
-                    ->counts('invites')
-                    ->sortable(),
-                // Tables\Columns\TextColumn::make('xlsforms_count')
-                //     ->label('# Xlsforms')
-                //     ->counts('xlsforms')
-                //     ->sortable(),
-                Tables\Columns\TextColumn::make('created_at')
-                    ->sortable(),
-            ]);
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ListTeams::route('/'),
-            'create' => Pages\CreateTeam::route('/create'),
-            'edit' => Pages\EditTeam::route('/{record}/edit'),
-            'view' => Pages\ViewTeam::route('/{record}'),
-        ];
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            UsersRelationManager::class,
-            InvitesRelationManager::class,
-            // XlsformsRelationManager::class,
-        ];
-    }
-
-    public static function infolist(Infolist $infolist): Infolist
-    {
-        return $infolist
-            ->schema([
-                TextEntry::make('description')->hiddenLabel(),
-            ]);
-    }
-}
