--- src/Filament/Admin/Resources/TeamResource/Pages/CreateTeam.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource\Pages;
-
-use Filament\Resources\Pages\CreateRecord;
-use Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource;
-
-class CreateTeam extends CreateRecord
-{
-    protected static string $resource = TeamResource::class;
-}
