+++ src/Filament/Admin/Resources/Teams/Pages/ViewTeam.php
@@ -13,7 +13,7 @@ class ViewTeam extends ViewRecord
 {
     protected static string $resource = TeamResource::class;
 
-    public function getTitle(): string | Htmlable
+    public function getTitle(): string|Htmlable
     {
         return $this->getRecord()->name;
     }
@@ -23,7 +23,7 @@ class ViewTeam extends ViewRecord
         return [
             Actions\EditAction::make(),
             Actions\DeleteAction::make()
-                ->modalDescription('WARNING: Please do not delete when there is actual survey data collected, as deletion is unreversable. Are you sure you would like to do this?'),
+                ->modalDescription('WARNING: This will permanently delete this team and all associated data. This action cannot be undone. Do you wish to continue?'),
         ];
     }
 }
