+++ src/Filament/Admin/Resources/Teams/RelationManagers/InvitesRelationManager.php
@@ -2,6 +2,11 @@
 
 namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\TeamResource\RelationManagers;
 
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\CreateAction;
+use Filament\Actions\DeleteAction;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
 use Filament\Resources\RelationManagers\RelationManager;
 use Filament\Tables;
 use Filament\Tables\Table;
@@ -9,18 +14,6 @@ use Filament\Tables\Table;
 class InvitesRelationManager extends RelationManager
 {
     protected static string $relationship = 'invites';
-
-    // we do not need a form here, team invitation form is defined in User resources as a headerAction
-    //
-    // public function form(Form $form): Form
-    // {
-    //     return $form
-    //         ->schema([
-    //             Forms\Components\TextInput::make('email')
-    //                 ->required()
-    //                 ->maxLength(255),
-    //         ]);
-    // }
 
     public function table(Table $table): Table
     {
@@ -48,15 +41,15 @@ class InvitesRelationManager extends RelationManager
                 //
             ])
             ->headerActions([
-                Tables\Actions\CreateAction::make(),
+                CreateAction::make(),
             ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-                Tables\Actions\DeleteAction::make(),
+            ->recordActions([
+                EditAction::make(),
+                DeleteAction::make(),
             ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
+            ->groupedBulkActions([
+                BulkActionGroup::make([
+                    DeleteBulkAction::make(),
                 ]),
             ]);
     }
