+++ src/Filament/Admin/Resources/Users/Tables/UserTable.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Stats4sd\FilamentTeamManagement\Filament\Admin\Resources\Users\Tables;
+
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+use Illuminate\Support\Str;
+
+class UserTable
+{
+    public static function configure(Table $table): Table
+    {
+        return $table
+            ->columns([
+                TextColumn::make('name')
+                    ->searchable(),
+                TextColumn::make('email')
+                    ->searchable(),
+                TextColumn::make('programs.name')
+                    ->searchable()
+                    ->badge()
+                    ->color('success')
+                    ->visible(config('filament-team-management.use_programs')),
+                TextColumn::make('teams.name')
+                    ->label(Str::ucfirst(Str::plural(config('filament-team-management.names.team'))))
+                    ->searchable()
+                    ->badge()
+                    ->color('success'),
+                TextColumn::make('roles.name')
+                    ->badge()
+                    ->searchable(),
+            ])
+            ->filters([
+                //
+            ])
+            ->recordActions([
+                EditAction::make(),
+            ])
+            ->groupedBulkActions([
+                BulkActionGroup::make([
+                    DeleteBulkAction::make(),
+                ]),
+            ]);
+    }
+}
